requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
numpy==2.1.3
pillow==11.0.0
requests==2.32.3
-e git+https://github.com/FifthGenML/twin_snakes.git@c511c8498e98b59f1d0a1ec14f32ec891194d20d#egg=twin_snakes
urllib3==2.2.3
